+++ src/App.tsx
@@ -5,9 +5,10 @@ import ProductPage from "./pages/product/ProductPage";
 import ProductDetailPage from "./pages/product/ProductDetailPage";
 import OrderPage from "./pages/order/OrderPage";
 import OrderCompletePage from "./pages/order/OrderCompletePage";
-import ProductManagePage from "./pages/admin/ProductManagePage";
 import CartPage from "./pages/CartPage";
 import ManagePage from "./pages/AdminProductPage";
+import ProductCreatePage from "./pages/admin/ProductCreatePage";
+import ProductEditPage from "./pages/admin/ProductEditPage";
 
 function App() {
   return (
@@ -19,8 +20,9 @@ function App() {
         <Route path="/order" element={<OrderPage />} />
         <Route path="/order/complete/:orderId" element={<OrderCompletePage />} />
         <Route path="/cart" element={<CartPage />} />
-        <Route path="/admin/products/:id" element={<ProductManagePage />} />
         <Route path="/admin/product" element={<ManagePage />} />
+        <Route path="/admin/product/new" element={<ProductCreatePage />} />
+        <Route path="/admin/product/:id" element={<ProductEditPage />} />
       </Route>
     </Routes>
   );
